worker/requirements.txt
Flask==1.1.1
gunicorn==19.9.0
google-cloud-tasks==1.3.0
googleapis-common-protos==1.6.0
python-dotenv